pi2_bedroom/requirements.txt
RPi.GPIO
Adafruit_DHT
paho-mqtt
pigpio
